--- src/convert/from_bdrhap2_to_h5mu/config.vsh.yaml
@@ -1,56 +1,0 @@
-functionality:
-  name: "from_bdrhap2_to_h5mu"
-  namespace: "convert"
-  description: |
-    Convert the output of a BD Rhapsody pipeline to a MuData h5 file.
-  authors:
-    - __merge__: /src/authors/dorien_roosen.yaml
-      roles: [ author, maintainer ]
-  argument_groups:
-    - name: Inputs
-      arguments:
-        - name: "--id"
-          type: string
-          description: "A sample ID."
-          example: my_id
-          required: true
-        - name: "--input"
-          alternatives: [-i]
-          type: file
-          description: The output h5mu of a BD Rhapsody workflow.
-          required: true
-          example: sample.h5mu
-    - name: Outputs
-      arguments:
-        - name: "--output"
-          alternatives: [-o]
-          direction: output
-          type: file
-          description: "Output h5mu file."
-          required: true
-          example: output.h5mu
-        - name: "--output_compression"
-          type: string
-          choices: ["gzip", "lzf"]
-          required: false
-          example: "gzip"
-  resources:
-    - type: python_script
-      path: script.py
-  test_resources:
-    - type: python_script
-      path: test.py
-    - path: /resources_test/bdrhap_5kjrt/processed2/sample.h5mu
-    
-platforms:
-  - type: docker
-    image: python:3.11-slim
-    setup:
-      - type: apt
-        packages:
-          - procps
-      - type: python
-        __merge__: /src/base/requirements/anndata_mudata.yaml
-  - type: nextflow
-    directives:
-      label: [lowmem, singlecpu]
