+++ src/convert/from_bdrhap_to_h5mu/test.py
@@ -5,7 +5,7 @@ import numpy as np
 
 ## VIASH START
 meta = {
-    "executable": "target/docker/convert/from_bdrhap2_to_h5mu/from_bdrhap2_to_h5mu",
+    "executable": "target/docker/convert/from_bdrhap_to_h5mu/from_bdrhap_to_h5mu",
     "resources_dir": "resources_test/bdrhap_5kjrt/processed2"
 }
 ## VIASH END
@@ -33,7 +33,7 @@ prot_adata = data.mod["prot"]
 
 # check whether correct feature types are detected
 assert np.array_equal(rna_adata.var["feature_type"].unique(), ["Gene Expression"]), "RNA expression should only contain Gene Expression vars."
-assert np.array_equal(rna_adata.var["reference_file"].unique(), ["reference_bd_rhapsody_v2.tar.gz"]), "Wrong reference file detected for Gene Expression vars."
+assert np.array_equal(rna_adata.var["reference_file"].unique(), ["reference_bd_rhapsody.tar.gz"]), "Wrong reference file detected for Gene Expression vars."
 assert "ADAMTSL4" in rna_adata.var_names, 'RNA modality should contain gene "ADAMTSL4".'
 assert np.array_equal(rna_adata.obs["library_id"].unique(), ["12ABC & 12WTA"]), "Gene Expression .obs library_id should equal '12ABC & 12WTA."
 # Enable after test resources have been synced with bd rhapsody processed data including smk reads
