+++ src/mapping/bd_rhapsody/rhapsody_pipeline_2.2.1_nodocker.cwl
