+++ src/reference/build_bdrhap_reference/run_test.sh
@@ -13,24 +13,15 @@ function clean_up {
 }
 trap clean_up EXIT
 
-function seqkit_head {
-  input="$1"
-  output="$2"
-  if [[ ! -f "$output" ]]; then
-    echo "> Processing `basename $input`"
-    seqkit subseq -r 1:50000 "$input" | gzip > "$output"
-  fi
-}
-
-seqkit_head "$meta_resources_dir/reference_gencodev41_chr1/reference.fa.gz" "$tmpdir/reference_small.fa.gz"
-zcat "$meta_resources_dir/reference_gencodev41_chr1/reference.gtf.gz" | awk '$4 < 50001 {print ;}' | gzip > "$tmpdir/reference_small.gtf.gz"
+seqkit subseq -r 1:50000 "$meta_resources_dir/reference.fa.gz" | gzip > "$tmpdir/reference_small.fa.gz"
+zcat "$meta_resources_dir/reference.gtf.gz" | awk '$4 < 50001 {print ;}' | gzip > "$tmpdir/reference_small.gtf.gz"
 
 
 echo "> Running $meta_functionality_name, writing to $tmpdir."
 $meta_executable \
   --genome_fasta "$tmpdir/reference_small.fa.gz" \
-  --transcriptome_gtf "$tmpdir/reference_small.gtf.gz" \
-  --output "$tmpdir/myreference.tar.gz" \
+  --gtf "$tmpdir/reference_small.gtf.gz" \
+  --reference_archive "$tmpdir/myreference.tar.gz" \
   ---cpus 2
 
 exit_code=$?
