+++ src/workflows/ingestion/bd_rhapsody/test.nf
@@ -11,16 +11,15 @@ workflow test_wf {
     [
       [
         id: "foo",
-        mode: "wta",
-        input: file("$resources_test/bdrhap_5kjrt/raw/12WTA*.fastq.gz"),
-        reference: resources_test.resolve("reference_gencodev41_chr1/reference_bd_rhapsody.tar.gz"),
-        transcriptome_annotation: resources_test.resolve("reference_gencodev41_chr1/reference.gtf.gz"),
-        putative_cell_call: "mRNA",
+        reads: file("$resources_test/bdrhap_5kjrt/raw/12*.fastq.gz"),
+        reference_archive: resources_test.resolve("reference_gencodev41_chr1/reference_bd_rhapsody.tar.gz"),
+        abseq_reference: resources_test.resolve("bdrhap_5kjrt/raw/BDAbSeq_ImmuneDiscoveryPanel.fasta"),
+        cell_calling_data: "mRNA",
         exact_cell_count: 4900
       ] 
     ])
     | map{ state -> [state.id, state] }
-    | bd_rhapsody 
+    | bd_rhapsody
     | view { output ->
       assert output.size() == 2 : "outputs should contain two elements; [id, file]"
 
@@ -29,13 +28,13 @@ workflow test_wf {
 
       assert id == "foo" : "Output ID should be same as input ID"
       assert "output_raw" in data : "Output should contain output_raw"
-      assert "output_h5mu" in data : "Output should contain output_h5mu"
-      assert data.output_h5mu.toString().endsWith(".h5mu") : "Output file should be a h5mu file. Found: ${output[1]}"
+      assert "output" in data : "Output should contain output_h5mu"
+      assert data.output.toString().endsWith(".h5mu") : "Output file should be a h5mu file. Found: ${output[1]}"
       "Output: $output"
     }
 
     | bd_rhapsody_test.run(
-      fromState: ["input": "output_h5mu"]
+      fromState: ["input": "output"]
     )
 
     | toList()
