--- src/workflows/ingestion/bd_rhapsody2/nextflow.config
@@ -1,10 +1,0 @@
-manifest {
-  nextflowVersion = '!>=20.12.1-edge'
-}
-
-params {
-  rootDir = java.nio.file.Paths.get("$projectDir/../../../../").toAbsolutePath().normalize().toString()
-}
-
-// include common settings
-includeConfig("${params.rootDir}/src/workflows/utils/labels.config")
