--- src/workflows/test_workflows/ingestion/bd_rhapsody2/config.vsh.yaml
@@ -1,34 +1,0 @@
-functionality:
-  name: "bd_rhapsody2_test"
-  namespace: "test_workflows/ingestion"
-  description: "This component test the output of the integration test of the bd_rhapsody workflow."
-  authors:
-    - __merge__: /src/authors/jakub_majercik.yaml
-  argument_groups:
-    - name: Inputs
-      arguments:
-        - name: "--input"
-          type: file
-          required: true
-          description: Path to h5mu output.
-          example: foo.final.h5mu
-  resources:
-    - type: python_script
-      path: script.py
-    - path: /src/utils/setup_logger.py
-    - path: /src/base/openpipelinetestutils
-      dest: openpipelinetestutils
-platforms:
-  - type: docker
-    image: python:3.12-slim
-    setup:
-      - type: docker
-        copy: ["openpipelinetestutils /opt/openpipelinetestutils"]
-      - type: apt
-        packages: 
-          - procps
-      - type: python
-        packages: /opt/openpipelinetestutils
-      - type: python
-        __merge__: [/src/base/requirements/anndata_mudata.yaml, /src/base/requirements/viashpy.yaml, .]
-  - type: nextflow
